FIX: Expressions in definitions are split into several items on contained whitespace
By using eval in the "set --", escaped whitespace or quoting of the entire expression will avoid the split.
The whitespace is also problematic in the extractions of individual :-separated fields; need to quote the variables there.
Mention the escaping / quoting in the documentation and adapt the existing custom config item.