Add special handling and mentioning of a custom:false:... item that always runs
For some customizations, it may be hard to detect whether they have already run or not. Storing a marker flag is always a possibility, but that then precludes the use of a simple inline expression, and a custom script has to be provided. For cheap actions that don't do harm if run again, just relying on the user's answer and its memoization may be enough.