custom type: ENH: Allow command-line arguments after executable command action
Especially in combination with sudo it can be advantageous to be able to pass arguments to the custom command in the directory tree (e.g. $USER is not the original one after sudo, and environment variables are not forwarded by sudo by default).
When the action is reused in the CHECK part, the command-line arguments are stripped off.