BUG: pre-/postinstall actions are not found
As the --pre and --post command-line arguments are executed on the fly, the {pre,post}installActionsDirspec is not yet defined because it now depends on a possibly given --base-dir argument.
As these had been shoehorned in (via the clumsy prePostStatus variable), switch to collecting the commands in a specialHandlingCommands array which is then interpreted _after_ command-line argument parsing. The only downside here is in the necessary quoting. (Another option would have been an updateBaseDir() function that gets invoked before command-line argument parsing and then on each --base-dir argument, but that also isn't clean and introduces another temporal dependency.)