Minor: Add separate subject for NPM
So that it's addressed via the same subject as in the JavaScript definition group. Let's keep it there, though, because for JavaScript development, we want NPM unconditionally, not just as an implicit dependency of installing an NPM package. (Similarly, pip3 would likely be added to a programming/Python definition group.)