Exclude require:, {pre,post}install: items from the (missing) item count
The user is interested in the actual software artifacts that are getting installed. As pre- and postinstall actions depend on the selection of the main item (and wouldn't run on their own), they should not be counted there. With the introduction of the require: items, the tally got distorted even more.