ENH: Add thunderbird type
Basically a copy of the firefox type, with s/Firefox/Thunderbird/ and adaptation of the profile dir.
Since Thunderbird cannot be launched with add-on URLs, and add-ons apparently either need to be searched from within Thunderbird or downloaded in a browser and installed from file, we just launch those URLs in the default browser here (with a small appended shell comment for which profile this is).