FIX: Ensure that the missing item that triggers pre-/postinstall actually is the previous one
And not just the previous missing one, for example with
    A B postinstall:X
Do not run X if A is missing, only if B is missing.
Record lastRealItem in addition to lastMissingItem, and only add the pre-/postinstall if both are equal.
No last real item is a syntax error that now can be detected and do an abort.