ENH: Postinstall execution is only counted as run once if it was actually executed and succeeded
The use of "once" doesn't fix our use case very well, as this is more concerned about preventing concurrent executions, and therefore records before execution starts. We rather need to only record after a successful execution, and have the additional separation of the check whether the postinstall-or-once: applies, and the actual execution that is only triggered later by the user (or maybe not). Already setting the "ran once" flag just when the command is printed to the user runs the risk of that this action is never actually performed.

Therefore, reimplement once ourselves (through onceMarker); it's just a thin wrapper around executionMarker, anyway. During the check phase, just do a query, and postpone the recording to the recursive invocation by the user. To be able to do the recording in addition to the execution of the postinstall action, we clone the --post argument into --post-or-once, and pass an additional --id ID for onceMarker, which is invoked if runPostInstallCheck() succeeds.