setup-software: UI: Highlight the type prefixes in the config help
Especially when listing all available types, the output is a bit messy. Instead of forcing every type to implement the highlighting, bank on the fact that each type's documentation starts with the prefix, and highlight that via an sed stage (that only activates when the pager is active, so captured output doesn't have it).