custom: Maintain the original order of custom actions
The associative arrays prevent the duplicate inclusion of the same action, but unfortunately the original order gets lost. For many other types that are passed en bloc to the respective package manager, this doesn't matter, and sorting order is only a cosmetic thing for the user, but here the execution order may matter, as multiple items in a single definition or definitions within the same group may depend on a certain order.
(I encountered the problem when setup-vncserver-autostart was executed before setup-vncserver, and the former requires setting a password which the latter does.)