ENH: Make --all also ask whether group directories should be considered
The desktop directory has gotten quite large, with 20 individual groups in it; going through all of them when we don't want any desktop stuff is tedious.
Include the group directories in the query and add a special check for them; if declined, all contained definition groups (and subdirs) are filtered from further iterations.