setup-software: BUG: Don't install a package twice if it's contained in multiple definitions wasn't fully applied
By using a addedNativePackages dictionary instead of array to store the selected packages and checking its keys in hasNative(), we can also detect packages that have already been selected (by another definition) in this installation session. Therefore, the application's logic will treat this item as already installed, and not add it (again).
We had added the key lookup in hasNative() without changing the variable type from array to dictionary.