Return 99 instead of 0 if getInstalled*() check in has*() fails to obtain the installed packages
The idea was that if the list of installed packages could not be obtained, we would assume that the definition is already there, so the user is not prompted. However, the actual implementations so far either do an "exit 3" if this cannot be obtained, or just ignore such failures. Also, with postinstall-or-once:, this "lying" doesn't work so well any longer, because it would still trigger the action.
Therefore, introduce a new 99 status that signals that the list of installed packages could not be determined, and we don't know whether the item is installed or not.
For the pre-/postinstall handling in hasOrAddItems(), we ignore 99, treating it like success so far - nothing changes.
For postinstall-or-once:, we don't show it once and also don't add this as a postinstall item.
In the regular loop, if one item's state cannot not be determined, the entire definition should be skipped, as we cannot ensure the correct installation.