wine+url: Refactoring: Do the processor architecture check only once during sourcing and define no-op function stubs then
Trying out an approach that can later be used for other native package management implementations. We still need the type registered, but any attempt to use it should cause the item (here: the entire definition) to be skipped.
We don't need to implement addWine(), because the negative hasWine() stub will never allow a wine item to be added.