yum-repo: Rename to yum+url, as this cannot just be used for installing third-party repositories
So this is quite similar to the existing deb+url: type. As the URL is handled by Yum itself, we don't have the deb-download-installer indirection, and therefore no caching and aging. As Yum for me is just a subsidiary package manager for cloud systems, that should be fine for now.