custom: ENH: Support CHECK:ITEM
If we want to install something based on a custom check, so far only require:CHECK could be used, but that affects the entire definition.
The custom type could only run executable commands or expressions, but not another item.
Especially because the custom type runs late, the delegation to add*() cannot be done in installCustom(); it would be too late.
The handling of the ITEM has be to done in addCustom(), and therefore requires a bit of duplicated parsing.
To avoid that an ITEM action is processed again in installCustom(), a new dict itemActions stores those.