{npm,pip3,snap}: A missing package tool means that no packages have been installed yet
Now that we check for the success of the package query, we have to explicitly handle the case where the package tool isn't installed yet. Assuming that packages can only be obtained through that package tool, we can safely deduce from the non-existence that no packages exist yet.
Once an item has been selected, the user will be prompted to install the package tool, anyway.