require: ENH: Allow negated require exectable command, too
Unlike with custom:, it may make sense to use a require executable command both with its original status and the negation, for example to have alternative definitions for Ubuntu and non-Ubuntu. One could use the shell negation in a REQUIREMENT-EXPRESSION, but we would lose the automatic invocation in the ./etc/require/ directory tree (so both require branches would either look different or both have to use an ugly long path).