{deb,wine}+url: Renaming: packageName -> applicationName
And remove the "package" from some other variables. In the deb+url type, packageName refers to PACKAGE[:...] in addDebUrl(), but to the [[SUBDIR/]NAME/] part in other functions, which is very confusing. Use applicationName for the latter; it's also more correct.
Apply the same renamings for consistency to the wine+url type as well.