Only use mkdir if necessary; don't use sudo for the $archiveDownloadInstallerCommand; instead, pass new --sudo argument
Running the entire $archiveDownloadInstallerCommand is a bad idea; the home directory is different, so a mounted install share may not be found / a different local /root/install directory would be created. The downloaded archive's right may complicate sharing. The root user might not have the HTTP proxy set up to download from the Internet.
Instead, container-download-installer now has a --sudo command that runs the unarchiving as root if necessary.
As we need to do checks for the $extractionDirspec now, anyway, omit the mkdir command if the directory already exists.