setup-software: BUG: Do not use $setupName as the picoDB namespace
The build dependencies (the only use case so far) are recorded for the entire system, across different setups, so they should not be scoped in separate namespaces.
The fact that installDebBuild() does not pass the --name "$setupName" to the recursive invocation caused a mismatch in lookup and persistence, and so the setup-vimdev has been continuously chosen for installation.