setup-software: Differentiate (and document) exit status
Exit status 1 is associated with failure of pre-/postinstall commands, and may also be a common value of an actual installation failure (with --execute).
To differentiate no action taken, use 4 and return that if no definition groups, no definitions were selected or to be installed.
Use 99 to show that there were no definitions to choose from.

FIX: As the definitions array isn't manipulated later, the message on it being empty should be moved up, so that the message is actually shown.