yum: Robustness: Do not block endlessly on repoquery when another yum execution is under way
Instead, check for yum's PID file and then limit the repoquery execution with a timeout of 2 seconds, so that the waiting message is shown just once.
Use timeout's exit status of 124 to detect the blocking and then completely abort the setup-software run, as likely several native packages will be affected and it doesn't make sense to retry this (with the corresponding 2-second delay) for each package.