custom: ENH: Also allow inverted check for EXECUTABLE-NAME or GLOB
The main use case was only running a custom action if something is there (to configure / modify it), but it might also make sense to run it if something is _not_ yet there (i.e. to install it).
Granted, all of that (negated and unnegated) can already be done via a TEST-EXPRESSION, but the motivation is in better readability and less fluff.