Support multiple definition group filters on a single line
Doing similar parsing as for the definitions (just skipping the extraction of a potential trailing comment; there's no use for that).
Add some more error handling as now definition group filters could be mixed with ordinary items; lastDefinitionGroupFilterItem serves both as a flag whether we've already seen one and stores the name for a better error message.