ENH: Allow suppression of entire definition group subdirs via filtering, too
The problem with directories is that there's nowhere to put the definition group filter definitions in :-) So let's check a special hidden file .groupdir-filter inside the subdir; we can reuse the same isDefinitionGroupCompletelyFiltered() for that, as long as we require that the configurator puts a dummy non-group filter in there, too. Then we can simply use the number of parsed definitions to see whether everything has been filtered or not.