BUG: "require.sh: line 50: 1: parameter null or not set" with "require:! true"
The combination of ! (which gets trimmed) and a following whitespace leads to a lookup with an empty requirement (that is expected to fail, but not like that).
Allow empty arguments and return (fail) early.
Apply the same precaution to the custom type, although I'm not sure that the problem could be triggered there at all.