Refactoring: Remove unnecessary eval
As long as the invoked variable is (part of) a function name, we don't need eval, as no word-splitting is involved.
Let's add double-quoting around them to ensure that the variable does not sneakily introduce additional arguments (so with typeFunction='Foo bar', it would fail to invoke "addFoo bar" "$name" instead of invoking addFoo bar "$name").