CHG: Combine multiple definition group filter in a single line with logical OR
Though much can be handled flexibly (and in a more straightforward way) with a REQUIREMENT-EXPRESSION, we may want to combine different types of definition group filters (e.g. group-tryout: and group-require:), or refer to built-in logic (group-require:native:vncserver group-require:hasDesktopEnvironment).
As definition group filters cannot be combined with other items, anyway, we can define the logical combination in any way we want. As logical AND can be trivially modeled by separate lines (any failing line will exclude the remainder of the group), let's define multiple filters in a single line as logical OR. The only downside is in the inconsistency with (non-group) filters, which are combined with AND (any single failing one will exclude the definition).