yum: Also run {pre,post}install actions for yum-utils
As we require the repoquery command right now, we unfortunately cannot use the isAvailableOrUserAcceptsNative() functionality (like in ppa.sh); adding a preinstall:native:yum-utils would again trigger the check whether it is already installed; endless recursion. So we have to directly use the "yum install" there.
Problem is that we may have a bad Yum proxy config; my setup-cloud-software has a custom check for that, but it comes too late; the yum-utils installation is already hanging.
Now that preinstall actions are run prior to the yum-utils installation, we have a chance in setup-cloud-software to define a etc/preinstall/yum-utils/yum-proxy.sh script that silently corrects this.