Refactoring: Make messagePrintf() still observe $isQuiet flag and change the isAvailable*() implementations to temporarily set the flag
Redirecting stderr to the null device means that yum.sh's haveRepoquery() cannot redirect the "yum install yum-utils" output to stderr to still display it. When this is invoked via a require:yum:... item, isAvailableYum() is invoked, in turn invokes hasYum(), and eventually haveRepoquery() may need to install the yum-utils package first. We do want to show those steps to the user (as they take some time and modify the system); it's just that require actions don't want the usual "Failed to obtain package list" error that has*() to be shown.
We could solve this by passing a new isShowMessage flag into has*() and set it in isAvailable*(), but as has*() is also directly invoked and takes arbitrary arguments already, that would be complex. Instead, make use of the existing $isQuiet flag by making messagePrintf() observe it (usually messagePrintf() gets overwritten with a no-op stub if --quiet is passed), and set that locally.