Only use regular items (from the typeRegistry) for the subject fallback
Filters (e.g. require:, tryout:) do not belong to the actual definition (and may change and then mess up an existing memoization).
Likewise, preinstall:, config:, etc. offer very little additional information and often are very technical (and long), so omit them as well.
I don't think it's possible to have a valid definition without regular items, but just in case, fall back to all concatenated items (as before).
The only inconsistency that this creates is that the plural is still based on the full item count.