Refactoring: Remove {custom,require}ActionsDirspecs from the main script
As these are only used by the corresponding types. The algorithm becomes a little bit less efficient now, as non-existing subdirs in base directories are checked for the full custom action / requirement, but as the number of base directories is likely very low (usually 1 or 2), this should not matter. We could instead build the original lists inside the types, but it could not be done on sourcing (as the base directories are not yet parsed then), so it would have to be done on-demand.