setup-software: ENH: Add built-in decorate: type
This is somewhat similar to the config: type in that it affects the preceding item, and influences the generated install command. But it does so by prepending a CMDPREFIX to the install command.
This requires the cooperation by individual types; the new submitInstallCommand() function offers some base functionality and can replace the existing toBeInstalledCommands+=("$cmd").