config: Add support for decorate: type annotations
The decorations are accessed through the full item; this type has just stored the customAction part so far, as that is all that is required to generate the install command. We can cram the full item into the value of the addedCustomActions dict, to avoid the definition of another dict just for that.