install: Refactoring: Don't store the install command with --check in addedInstallActions
So far, there was only one parse in hasInstall() and two (with and without --check) in addInstall(); now it's the other way around. Considering that additions are less frequent (at most equal) as checks, this is more inefficient, but it shouldn't matter, and on the positive side this distinction between addedInstallActions and addedInstallActionList is removed.