setup-software: CHG: Put config: items before instead of after the real item
So that the configuration already is available to the item's add*() function, not just install*(). An implementation that is able to already prepare different install lists can be simpler than having to collect everything in one list during add*(), and then filter that list into multiple during install*() based on the (only now available) configuration.
As the config items are only fully assembled after the real item, we already know whether the real item is missing or not, so that check in addConfig() can go. (And actually has to go, because addConfig() is now invoked _before_ add*().)
A separate missingItemConfigurations is used so that these can be passed before missingItems to addItems(), so that the configurations are already defined when the real item is added. As a bonus, the config: item itself isn't tallied any longer in the missing item count - as it's non-actionable meta data, it shouldn't.
Handle config: similar to decorate:; however, we cannot simply invoke addConfig() directly, but have to build up missingItemConfigurations, and do the processing only when the user actually accepted the definition.