snap: ENH: Support classic mode installations via configuration
Some snaps (e.g. kubectl) do not work in security confinement and need to be installed via "snap install --classic". Allow that via an item configuration.
As the configuration now already is available in addSnap(), we can simply duplicate the addedSnapPackages array into addedSnapClassicPackages, and process both separately. This is easier than having to filter a single list in installSnap() based on the configuration.