FIX: addPreinstall() arguments are swapped and is[Dependency]AvailableOrUserAcceptsNative() needs to skip the addedItems check of addPreinstall()
ITEM is the first one, followed by ACTION; I probably got confused by the concatenation in a single argument; that's why I first extracted addPreinstallArgs() to make it clearer.
Only postinstall evaluates the ITEM (to check whether it needs to run); preinstall runs unconditionally, so ITEM is ignored. What needs to run is the ACTION, which in the case of isAvailableOrUserAcceptsNative() is the native:${packageManagerPackageName}. However, addPreinstall() checks whether the ITEM has been added at all, but that's not desired when we misuse the preinstall logic to run special install actions in isAvailableOrUserAcceptsNative() / isDependencyAvailableOrUserAcceptsNative(). Therefore, directly invoke addPreinstallArgs(); with this, we also don't need to concatenate the ITEM and ACTION args.