{firefox,thunderbird}: Check for required jq dependency and ask for (immediate) installation if it's not yet installed
We do have the dependency in my personal software definitions (both in programming-common and Unixhome); I could demand that the latter is installed first, as setup-software is part of the Unixhome, but this is clumsy. It's more comfortable to directly ask and on the spot install the missing dependency on the first use of the type, like we do in the yum type. (But here we can simply rely on the native package manager to install the dependency for us.)