FIX: Also check for flatpak dependency in flatpak-remote:, and clarify the relationship with flatpak:
We might (theoretically) just install a flatpak-remote without any Flatpak package; this would then miss the native flatpak package dependency.
Unfortunately, if we add the isAvailableOrUserAcceptsNative() check, the flatpak dependency will be installed twice if there's also a missing flatpak: item. This is because the preinstall: invokes a child process, and the selection of the flatpak dependency there does not get communicated back to the parent script; nor is there a good way to do this.
Though I've recommended the preinstall with ppa: (together with apt:), let's deviate from this here, as this little inconsistency avoids a huge implementation effort in the main script and types.
Without preinstall:, in order for the native flatpak dependency to be issued before the flatpak-remote:, we need to increase its priority to come after the native types. For ppa:, it had to come before, but here it's different. So let's choose subsequent numbers 201 + 202.