ENH: Add invokeCheck() function for use in types to wrap potentially long-running checks
Checks may have to do stuff that can take some time. When a selection or group has already been passed, this delay happens before any user interaction, which is bad.
Use my invocationMessage to temporarily show a note if the check takes more than half a second.
The invokeCheck() function replaces the (single-arg) eval in the custom and require types.

The require type so far passed a double-quoted shell variable to eval, as the command-line now is interpreted by the external invocationNotification tool, that's not possible any longer. Instead of exporting the variable (which would be an ugly exposure of internals), interpolate the variable contents, which must now be quoted. Modify getRequirementExecutable() to do that (and rename as getQuotedRequirementExecutable()).