tar+url: Force owner to be the current user
According to the man page, extracting as yourself is the default for ordinary users, but not root. (unzip needs a -X option to set embedded ownership information (which is rarely contained in ZIP archives, anyway).)
Some archives are packed with funny users (e.g. Eclipse); default to root when extracting system-wide.