special-install: ENH: Also support SOURCE-FILE DESKTOP-ENTRY syntax for copying from existing source
When the whole installer is in the install repo, startmenu+url can be used, but for "online" installations it's better to put extra files into the setup definition. Another use case might be that the installer ships with an desktop entry, but not in a location that's active; then, a copy with an absolute source makes sense.