CHG: Rename {startmenu,icon}[+url] to user* and add system-wide variants
All of the other types default to system-wide installations, so the default for icons and start menu entries should be system-wide, too.
There's still a use case for user-specific start menu entries and icons, though; some applications are only available to me.
Add separate types (with user* prefix) instead of adding --user or --system-wide command-line arguments; that wouldn't look so nice and complicate the parsing.
For the implementation, all we have to do is pass the new --system-wide command-line flag to the corresponding commands.
Apparently, there's no system-wide autostart location (as autostart should be a personal decision), so keep the default "autostart" type and yet keep it user-scoped.