BUG: A single group is still queried even through the entire group is filtered
Because that group is not added to the filteredGroups array, as we're continuing with the next group if there's only one contained group.
Move the condition for a single group around the $obtainGroups itself, so that the query is still skipped, but the single group is still added to filteredGroups if it's completely filtered.