custom: CHG: Turn special "false" check into "once" and memoize a successful execution
Because getting asked about the installation over and over again is tedious. The initial intention to delegate this to the user (who hopefully remembers whether he did that already in the past) is sabotaged by choosing a persistent option (like "yes" or "always") - the user won't be asked again.
For a notification or shell command, we can append the memoization via &&.
For an ITEM action, we synthesize a postinstall action that executes the memoization if the ITEM got installed.