dummy: ENH: Allow dummy:NAME:ITEM to act as references for a real ITEM, allowing dynamic require: of those
Several programs used to be installed as native (apt) packages, but are only delivered as snaps in recent Ubuntu distributions: Firefox, Ubuntu Make, etc. That's modeled as two alternative definitions with a version require: before each.
When these have many custom and postinstall actions, we'd have to duplicate all of those, which isn't nice. It would be great if we could attach a unique reference to the varying item, and then require that in a second (version-independent) definition with those custom / postinstall actions.
The dummy item is a perfect place to implement it. When given both NAME and ITEM, it will indicate itself as available if the ITEM is already installed, so other require: will pass. It ITEM isn't installed yet, it will indicate itself as not available, so if the user selects the definition, it'll get added, ensures the ITEM actually got added, too, and then marks itself as available, so following definitions can require: it.