setup-software: Cache positive answer from memoizeDecision and don't query again within the session for the same subject
When using dummy references to link together different require: items, a definition is broken into multiple definitions (as require:'s scope is for the entire definition). But it's odd for the user to be asked twice about the same definition - that happens when memoizeDecision is instructed to _not_ memoize ([c]onfirm each).
Add an in-memory cache for the current session, and bypass memoizeDecision if the answer is already found in the cache.