Add --post-context and addPostinstallContextCommand() to signal to postinstall actions that the dummy has been installed
If a dummy: item reference is used to split a single definition into many (so that different require: can be used on alternatives), and the common part uses postinstall[-or-once]:, it'll complain about "Skipping postinstall action, because installation of dummy:... was not successful." That's because the dummy installation is a no-op, and outside of the original invocation there's no record of whether the dummy was active.
Define a new --post-context, similar to --post, to transfer an arbitrary context to the post-install invocation.
The dummy: type uses that to transfer its addedDummyPackages hash in (previously empty) installDummy(), received by new setInstalledDummyPackages().
The --post-context arguments are prepended and additionally eval'd immediately; this (doubly) ensures that the context is set when a --post action is executed. The faked existence of the dummy package means that the postinstall action will now run. Post-context arguments are stored separate from addedPostinstallArgs, and a recursive invocation will only happen if the latter is non-empty, this ensures that a bare context won't trigger a recursive invocation (that would result in a usage error, as there's no action to perform) on its own.