Add apt-reinstall variant of apt-remove
We cannot simply use apt-remove:FOO apt:FOO, as the check for missing items runs over the entire definition, and only based on that static selection of missing packages will stuff be added. Now this could be solved by distributing the items over multiple eponymous definitions linked by a dummy item, but that's clumsy and not very readable.
apt-reinstall works like apt-remove, but also schedules the package for addition; the typeInstallOrder then ensures that those run in the right order.