BUG: postinstall-or-once drops a custom ITEM prefix
Most of my postinstall-or-once: definitions have been with EXECUTABLE-COMMAND or TEXT-FILE, but if an ITEM:... is used, that ITEM prefix needs to be kept so that the ITEM is recognized and properly executed.
> $ /home/inkarkat/Unixhome/bin/ingo/software/bin/setup-software --execute --add-base-dir /home/inkarkat/Unixhome/bin/setups/personal-software/etc --post-or-once imagemagick --action --sudo\ --directory\ /\ --strip\ 0\ imaging/imagemagick/policy.xml.patch --id patch
> /home/inkarkat/Unixhome/bin/ingo/software/bin/setup-software: line 1000: eval: --: invalid option
> eval: usage: eval [arg ...]

The problem is in the optional [ID:]; an ITEM prefix is mistaken for that.
Use the type registry to disambiguate; if the prefix is found in there, it's taken as an ITEM without an ID; else, it's a custom ID.