deb+url: ENH: Also do a package name-only comparison of added records
It may happen that the same package is defined elsewhere differently (and a slight discrepancy like in the MAX-AGE would already disrupt the comparision). As the package name is the only thing that count (i.e. finally gets installed), do a check for just that, too.
This also allows dummy items to refer to a deb+url via just deb+url:PACKAGE:*:, i.e. not needing to repeat the entire item.