ENH: Add update-software-done and print that at the end of an actual update run
If the update succeeds until the end, it can be assumed (without checking again) that all updates have been installed, and the MOTD update-available record can be cleared.
We do that by appending a custom command at the end of the update commands; it determines the updater name from the parent command name, and then removes the corresponding entries from the shared updates-available record.