FIX: Error from parseGitrepo() does not abort the script
Because there's just nothing to eval; the exit status of the command interpolation is lost. We could echo "exit 3", but that would force the flow control onto the caller. It's cleaner (though a bit cumbersome) to first assign the command substitution's result to a parse variable and check the exit status there, and only then do the eval.