ENH: Support command substitution in "install:'<(command) DEST-FILE'"
Useful e.g. for Bash completions that are produced by a command (kubectl and khal do this).
We cannot use the parsed $sourceFilespec in the install command - this is the /dev/fd/63 temporary file descriptor. Fortunately, the source command substitution is still available for us (prior to eval'ing it (which we only needed for the command-line argument validation); we just need to use that (without additional escaping) - but only if this actually is a command substitution (else if would break source files relative to a base dir via getAbsoluteOrFilesFilespec()).
Likewise, getAbsoluteOrBaseFilespec() needs an additional condition for file descriptors; this only existed during the eval for parsing, not when $sourceFile is checked for base dirs, but it will be re-created by the shell when the actual installation commands are run.