ENH: Also support config: items for the has*() check, and add support for custom: environment variables
For the check, the config is not yet committed, but only applicable to the check itself; therefore, a separate hasConfiguration dict is used, and that is cleared before each has*() invocation.
The custom type just includes the configured environment variables (flattened to one line) in the eval of the custom check / action.