update-software: Only use administrative rights to update the stamp file when needed (cp. ea92162, ffbc714)
> $ update-software-done update-personal-software
> sed: couldn't open temporary file /var/lib/update-notifier/sedak05gI: Permission denied

It actually isn't about writing to the stamp file itself (which setup-software-updates-available makes writable to the user), the problem is that the in-place sed modification fails because sed will always store the temp file in that directory, and the user has no access to it.
Use my new sedInplaceWithoutDirPermissions variant in update-software-done so that only write permissions to the stamp file itself are required.
Omit the sudo unless the current user has lost write access to the stamp file somehow.

(This relates to commit ea92162a5aa6ad40b1d3ab98e1df50990d1725d0 and ffbc714f3eebaeb3706c69d845987db41ecbf569)