FIX: Group descriptions go into the bag / the software update list
Any group description has to be redirected to stderr, as stdout is captured on regular use.
Furthermore, the group dir description is not controlled by $printGroups, but output unconditionally, which mistakenly includes the descriptions in the update-software output.
Change $printGroups to a $isPrintGroups flag; this also avoids an unnecessary invocation of withHeader when it's off.