firefox: Robustness: Read default profile name from Firefox's profiles.ini
The globbing apparently still has problems; a fresh setup of my Firefox add-ons extracted "default" instead of "default-release", and because the Firefox add-on command-line passes the profile (even when default), this led me to installing all add-ons into the wrong profile! (Such might have happened before; I'm having an eery deja-vu...)
It's more robust (when both default and default-release exist) to directly read the Firefox configuration. (The parsing however is simplistic, and just covers that one profile might have a "Default=1" as well.)