Adapt: npm is not a native package any longer
I had already adapted the programming/JavaScript definitions in personal-software, but failed to apply these to the npm bootstrapping here.
As we now need a full definition for npm, the isAvailableOrUserAcceptsNative() function doesn't cut it any longer; the previous refactorings allow to invoke parseDefinitionGroups() separately from the main loop, so an isAvailableOrUserAcceptsGroup() alternative can be added.
The new nodejs group is placed into lib/definitions, not etc/definitions, so that it isn't automatically activated. Fortunately, the definition group can be both name or filespec.