ENH: Add apt-upgrade: variant of apt-reinstall:
That doesn't actually remove the package, but simply attempts an upgrade (if available, e.g. triggered by a change in the apt repository so that a different version is selected, or if a different version is now pinned). The package manager itself performs regular updates; that's not part of this use case.