nodejs: Use apt-upgrade:
We don't need to uninstall the existing nodejs package after apt-repo: has installed the nodesource repo, after the package update the package manager will automatically select the new current version. However, we need an explicit package upgrade if npm: items are to be installed (likely the reason for the selection of the npm dependency); waiting for the package manager to pick this up on the next system upgrade isn't sufficient.