ENH: Add var: meta item that allows de-duplication of version numbers etc. in definitions
To have them function just like regular exported environment variables they need to be defined in the central eval() of the definition that splits it into items. Unfortunately, that poses a hen-and-egg problem, as we want the variables to be defined within that definition.
There's already a sort of pre-parsing for the definition filters. Let's hook onto those; by requiring the var: items to be first, we can evaluate them, then do a re-evaluation of the definition in case we've encountered some and shift them off. This means that variables cannot refer to each other, and cannot be defined just for following items, but that should be fine.
The alternative, implementing definition group-scoped variables, is also hard to implement as we're first parsing all definition groups into a single steam of definitions, so implementing the scoping to the group would be difficult.
After the definition, the variable definition needs to be undone (we cannot simply use a subshell for scoping as all the other script data structures still need to be updated). For that, we need to distinguish between previously undefined and defined (but possibly empty) environment variables. This is solved by storing the corresponding restoration command (either unset VARNAME or VARNAME=PREVIOUS_VALUE) in the variables dict.